Update app.py
locs dtype change 'uint8' -> 'uint16'

+++ flask_ai/app.py
@@ -83,7 +83,7 @@ def Processing():
 
   except:
     pass
-  locs = np.array(locs,dtype='uint8')
+  locs = np.array(locs,dtype='uint16')
   locs = locs.tolist()
   result = {'locs':locs, \
             'labels':labels}
